Revert "Merge branch 'master' of https://github.com/dbarf22/group27cinema"
This reverts commit 8f5444236760ba8603027baca49f5a7605d82555, reversing
changes made to eb114cb5e1694daa835844720a38165d3c11ce17.

+++ cinema_backend/src/main/java/com/group27/cinema_backend/dto/CheckoutResponse.java
@@ -1,6 +1,5 @@
 package com.group27.cinema_backend.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class CheckoutResponse {
@@ -8,7 +7,7 @@ public class CheckoutResponse {
     private boolean success;
     private String message;
     private Integer bookingId;
-    private BigDecimal totalPrice;
+    private double totalPrice;
     private List<Integer> bookedSeatIds;
 
     public boolean isSuccess() {
@@ -35,11 +34,11 @@ public class CheckoutResponse {
         this.bookingId = bookingId;
     }
 
-    public BigDecimal getTotalPrice() {
+    public double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(BigDecimal totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
